+++ packages/abstract-extractor/lib/ergo/network/ExplorerNetwork.ts
@@ -29,7 +29,8 @@ export class ExplorerNetwork {
         box.additionalRegisters,
         'serializedValue'
       ),
-      assets: box.assets?.map((asset) => pick(asset, ['tokenId', 'amount'])),
+      assets:
+        box.assets?.map((asset) => pick(asset, ['tokenId', 'amount'])) ?? [],
     };
   };
 
@@ -50,7 +51,8 @@ export class ExplorerNetwork {
         box.additionalRegisters,
         'serializedValue'
       ),
-      assets: box.assets?.map((asset) => pick(asset, ['tokenId', 'amount'])),
+      assets:
+        box.assets?.map((asset) => pick(asset, ['tokenId', 'amount'])) ?? [],
     };
   };
 
@@ -69,6 +71,7 @@ export class ExplorerNetwork {
         tx.dataInputs?.map((dataInput) => ({
           boxId: dataInput.boxId,
         })) ?? [],
+      // TODO: Add input extension to explorer local/ergo/rosen-bridge/scanner/-/issues/156
       inputs: tx.inputs?.map((input) => this.convertInputBox(input)) ?? [],
       outputs: tx.outputs?.map((output) => this.convertOutputBox(output)) ?? [],
     };
@@ -85,15 +88,16 @@ export class ExplorerNetwork {
         tx.dataInputs?.map((dataInput) => ({
           boxId: dataInput.id,
         })) ?? [],
+      // TODO: Add input extension local/ergo/rosen-bridge/scanner/-/issues/156
       inputs: tx.inputs?.map((input) => ({ boxId: input.id })) ?? [],
       outputs:
         tx.outputs?.map((output) => ({
           boxId: output.id,
           transactionId: output.txId,
           additionalRegisters: output.additionalRegisters,
-          assets: output.assets?.map((asset) =>
-            pick(asset, ['tokenId', 'amount'])
-          ),
+          assets:
+            output.assets?.map((asset) => pick(asset, ['tokenId', 'amount'])) ??
+            [],
           ergoTree: output.ergoTree,
           creationHeight: output.creationHeight,
           index: output.index,