+++ packages/observation-extractor/tests/extractor/cardano/koios.spec.ts
@@ -11,6 +11,7 @@ import { Buffer } from 'buffer';
 import { blake2b } from 'blakejs';
 import { ERGO_NATIVE_TOKEN } from '../../../lib/extractor/const';
 import { DataSource } from 'typeorm';
+import { TokenMap } from '@rosen-bridge/tokens';
 
 class CardanoKoiosExtractor extends CardanoKoiosObservationExtractor {}
 
@@ -32,9 +33,11 @@ describe('cardanoKoiosObservationExtractor', () => {
      *  field should fulfill expected values
      */
     it('should returns true valid rosen transaction', async () => {
+      const tokenMap = new TokenMap();
+      await tokenMap.updateConfigByJson(tokens);
       const extractor = new CardanoKoiosExtractor(
         dataSource,
-        tokens,
+        tokenMap,
         bankAddress
       );
       const Tx: KoiosTransaction = cardanoTxValid;
@@ -54,12 +57,12 @@ describe('cardanoKoiosObservationExtractor', () => {
         fromChain: 'cardano',
         toChain: 'ergo',
         fromAddress:
-          'addr_test1vzg07d2qp3xje0w77f982zkhqey50gjxrsdqh89yx8r7nasu97hr0',
+          'addr1qytsk73jatycajqksafza5z90cw3zj2exhtdqx226r2l6dphvyt647kn7zl3svpnzjmuty2sfsr28cmf3aaa263hazqqxwdedk',
         toAddress: '9i1EZHaRPTLajwJivCFpdoi65r7A8ZgJxVbMtxZ23W5Z2gDkKdM',
         height: 1,
-        amount: '10',
-        networkFee: '10000',
-        bridgeFee: '10000',
+        amount: '1635516886333',
+        networkFee: '3829872',
+        bridgeFee: '10376749',
         sourceChainTokenId:
           'ace7bcc2ce705679149746620de3a84660ce57573df54b5a096e39a2.7369676d61',
         targetChainTokenId: ERGO_NATIVE_TOKEN,
@@ -78,9 +81,11 @@ describe('cardanoKoiosObservationExtractor', () => {
      * Expected: processTransactions should returns true and database row count should be 0
      */
     it('database row count should be zero because of invalid bankAddress', async () => {
+      const tokenMap = new TokenMap();
+      await tokenMap.updateConfigByJson(tokens);
       const extractor = new CardanoKoiosExtractor(
         dataSource,
-        tokens,
+        tokenMap,
         'addr_test1qq5qeusgymq8ledv9gltp9fuh5jchetjeafha75n6dghur4gtzcgx'
       );
       const Tx: KoiosTransaction = cardanoTxValid;
@@ -101,17 +106,20 @@ describe('cardanoKoiosObservationExtractor', () => {
      * Expected: processTransactions should returns true and database row count should be 0
      */
     it('should returns false invalid rosen metadata', async () => {
+      const tokenMap = new TokenMap();
+      await tokenMap.updateConfigByJson(tokens);
       const extractor = new CardanoKoiosExtractor(
         dataSource,
-        tokens,
+        tokenMap,
         bankAddress
       );
       const Tx: KoiosTransaction = {
         ...cardanoTxValid,
-        metadata: {
-          '0': JSON.parse(
-            '{"to": "ergo","bridgeFee": "10000","toAddress": "9i1EZHaRPTLajwJivCFpdoi65r7A8ZgJxVbMtxZ23W5Z2gDkKdM","targetChainTokenId": "cardanoTokenId"}'
-          ),
+        auxiliary_data: {
+          prefer_alonzo_format: false,
+          metadata: {
+            '0': '{"map":[{"k":{"string":"to"},"v":{"string":"ergo"}},{"k":{"string":"bridgeFee"},"v":{"string":"10376749"}},{"k":{"string":"networkFee"},"v":{"string":"3829872"}},{"k":{"string":"toAddress"},"v":{"string":"9hZxV3YNSfbCqS6GEses7DhAVSatvaoNtdsiNvkimPGG2c8fzkG"}}}',
+          },
         },
       };
       const res = await extractor.processTransactions(