+++ packages/observation-extractors/bitcoin-observation-extractor/lib/BitcoinRpcObservationExtractor.ts
@@ -1,7 +1,7 @@
 import { AbstractObservationExtractor } from '@rosen-bridge/observation-extractor';
 import { BitcoinRpcTransaction } from '@rosen-bridge/bitcoin-rpc-scanner';
 import { BitcoinRpcRosenExtractor } from '@rosen-bridge/rosen-extractor';
-import { RosenTokens } from '@rosen-bridge/tokens';
+import { TokenMap } from '@rosen-bridge/tokens';
 import { AbstractLogger } from '@rosen-bridge/abstract-logger';
 import { DataSource } from 'typeorm';
 
@@ -11,7 +11,7 @@ export class BitcoinRpcObservationExtractor extends AbstractObservationExtractor
   constructor(
     lockAddress: string,
     dataSource: DataSource,
-    tokens: RosenTokens,
+    tokens: TokenMap,
     logger?: AbstractLogger
   ) {
     super(