+++ packages/scanner/lib/scanner/ergo/network/types.ts
@@ -3,8 +3,13 @@ export type NodeBlock = {
   transactions: Array<Transaction>;
 };
 
+export type InputExtension = {
+  [key: string]: string;
+};
+
 export type InputBox = {
   boxId: string;
+  extension?: InputExtension;
 };
 
 export type DataInput = {
@@ -30,8 +35,8 @@ export type OutputBox = {
   value: bigint;
   ergoTree: string;
   creationHeight: number;
-  assets?: Array<Asset>;
-  additionalRegisters?: AdditionalRegisters;
+  assets: Array<Asset>;
+  additionalRegisters: AdditionalRegisters;
   transactionId: string;
   index: number;
 };