+++ packages/scanners/bitcoin-rpc-scanner/lib/index.ts
@@ -1,2 +1,3 @@
 export * from './BitcoinRpcScanner';
+export * from './DogeRpcScanner';
 export * from './types';