+++ packages/watcher-data-extractor/lib/extractor/EventTriggerExtractor.ts
@@ -6,6 +6,7 @@ import { AbstractLogger } from '@rosen-bridge/abstract-logger';
 import {
   AbstractInitializableErgoExtractor,
   BlockInfo,
+  CallbackType,
   ErgoNetworkType,
   OutputBox,
   SpendInfo,
@@ -15,8 +16,12 @@ import EventTriggerAction from '../actions/EventTriggerAction';
 import { ExtractedEventTrigger } from '../interfaces/extractedEventTrigger';
 import { JsonBI } from '../utils';
 import { EventResult } from '../types';
+import EventTriggerEntity from '../entities/EventTriggerEntity';
 
-class EventTriggerExtractor extends AbstractInitializableErgoExtractor<ExtractedEventTrigger> {
+class EventTriggerExtractor extends AbstractInitializableErgoExtractor<
+  ExtractedEventTrigger,
+  EventTriggerEntity
+> {
   id: string;
   protected readonly actions: EventTriggerAction;
   private readonly eventTriggerErgoTree: string;
@@ -33,9 +38,10 @@ class EventTriggerExtractor extends AbstractInitializableErgoExtractor<Extracted
     RWT: string,
     permitAddress: string,
     fraudAddress: string,
-    logger?: AbstractLogger
+    logger?: AbstractLogger,
+    initialize = true
   ) {
-    super(type, url, address, logger);
+    super(type, url, address, logger, initialize);
     this.id = id;
     this.eventTriggerErgoTree = wasm.Address.from_base58(address)
       .to_ergo_tree()
@@ -123,7 +129,7 @@ class EventTriggerExtractor extends AbstractInitializableErgoExtractor<Extracted
         eventId: eventId,
         txId: box.transactionId,
         boxId: box.boxId,
-        boxSerialized: Buffer.from(parsedBox.sigma_serialize_bytes()).toString(
+        serialized: Buffer.from(parsedBox.sigma_serialize_bytes()).toString(
           'base64'
         ),
         toChain: Buffer.from(R5Serialized[2]).toString(),
@@ -201,9 +207,21 @@ class EventTriggerExtractor extends AbstractInitializableErgoExtractor<Extracted
             extras: [result, paymentTxId],
           });
       });
-      if (boxes.length > 0)
-        await this.actions.insertBoxes(boxes, block, this.getId());
-      await this.actions.spendBoxes(spendInfoArray, block, this.id);
+      if (boxes.length > 0) {
+        if (!(await this.actions.storeBoxes(boxes, block, this.getId()))) {
+          this.logger.warn(
+            `Data insertion failed at height ${block.height} for extractor ${this.id}`
+          );
+          return false;
+        }
+        this.triggerCallbacks(CallbackType.Insert, boxes);
+      }
+      const spentData = await this.actions.spendBoxes(
+        spendInfoArray,
+        block,
+        this.id
+      );
+      this.triggerCallbacks(CallbackType.Spend, spentData);
     } catch (e) {
       this.logger.error(
         `Error in storing data in ${this.getId()} of the block ${block}: ${e}`