+++ packages/abstract-extractor/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @rosen-bridge/abstract-extractor
 
+## 1.0.2
+
+### Patch Changes
+
+- Add input box extension to network api interfaces
+- Refactor query logic in abstract ergo extractor actions
+
+## 1.0.1
+
+### Patch Changes
+
+- Update node version to 20
+- Update logger package
+
+## 1.0.0
+
+### Major Changes
+
+- Move database action shared logic to abstract class
+- Add callback support to abstract ergo extractor and update abstract db action interface
+
+### Patch Changes
+
+- Ensure initialization with the node is stateful and resilient to network issues
+- Fix `processTransaction` to return false if data insertion encounters error
+- Modify the initialization procedure to retry upon failure
+
 ## 0.3.1
 
 ### Patch Changes