+++ packages/observation-extractors/bitcoin-observation-extractor/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @rosen-bridge/bitcoin-observation-extractor
 
+## 6.0.0
+
+### Major Changes
+
+- Update tokens package to v3.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @rosen-bridge/observation-extractor@7.0.0
+  - @rosen-bridge/bitcoin-esplora-scanner@1.1.1
+  - @rosen-bridge/bitcoin-rpc-scanner@1.1.1
+
+## 5.0.0
+
+### Major Changes
+
+- Change RosenTokens to TokenMap
+
+### Minor Changes
+
+- Add Doge RPC and Esplora observation extractor
+
+### Patch Changes
+
+- Update node version to 20
+- Updated dependencies
+  - @rosen-bridge/bitcoin-rpc-scanner@1.1.0
+  - @rosen-bridge/observation-extractor@6.0.0
+  - @rosen-bridge/bitcoin-esplora-scanner@1.1.0
+
 ## 4.0.10
 
 ### Patch Changes