+++ packages/observation-extractors/evm-observation-extractor/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @rosen-bridge/evm-observation-extractor
 
+## 5.0.0
+
+### Major Changes
+
+- Update tokens package to v3.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @rosen-bridge/observation-extractor@7.0.0
+  - @rosen-bridge/evm-rpc-scanner@2.0.3
+
+## 4.0.0
+
+### Major Changes
+
+- Change RosenTokens to TokenMap
+
+### Patch Changes
+
+- Update node version to 20
+- Updated dependencies
+  - @rosen-bridge/observation-extractor@6.0.0
+  - @rosen-bridge/evm-rpc-scanner@2.0.2
+
 ## 3.1.2
 
 ### Patch Changes