+++ packages/scanner/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @rosen-bridge/scanner
 
+## 5.1.0
+
+### Minor Changes
+
+- Update koios scanner to use the new serialized APIs
+
+### Patch Changes
+
+- Add input box extension to network api interfaces
+- Updated dependencies
+  - @rosen-bridge/abstract-extractor@1.0.2
+
+## 5.0.2
+
+### Patch Changes
+
+- Update node version to 20
+- Updated dependencies
+  - @rosen-bridge/abstract-extractor@1.0.1
+
+## 5.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @rosen-bridge/abstract-extractor@1.0.0
+
 ## 5.0.0
 
 ### Major Changes