+++ packages/scanners/bitcoin-rpc-scanner/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @rosen-bridge/bitcoin-rpc-scanner
 
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @rosen-bridge/scanner@5.1.0
+
+## 1.1.0
+
+### Minor Changes
+
+- Add Doge
+
+### Patch Changes
+
+- Update node version to 20
+- Updated dependencies
+  - @rosen-bridge/scanner@5.0.2
+
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @rosen-bridge/scanner@5.0.1
+
 ## 1.0.0
 
 ### Major Changes